+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/asn1/ObjectIdentifier.kt
@@ -1,5 +1,3 @@
-@file:OptIn(ExperimentalUnsignedTypes::class)
-
 package at.asitplus.crypto.datatypes.asn1
 
 import kotlinx.serialization.KSerializer
@@ -20,7 +18,6 @@ import kotlin.math.ceil
  * @param nodes OID Tree nodes passed in order (e.g. 1u, 2u, 96u, …)
  * @throws Asn1Exception if less than two nodes are supplied, the first node is >2 or the second node is >39
  */
-@ExperimentalUnsignedTypes
 @Serializable(with = ObjectIdSerializer::class)
 class ObjectIdentifier @Throws(Asn1Exception::class) constructor(@Transient vararg val nodes: UInt) :
     Asn1Encodable<Asn1Primitive> {
@@ -48,6 +45,10 @@ class ObjectIdentifier @Throws(Asn1Exception::class) constructor(@Transient vara
         if (other == null) return false
         if (other !is ObjectIdentifier) return false
         return nodes contentEquals other.nodes
+    }
+
+    override fun hashCode(): Int {
+        return nodes.hashCode()
     }
 
     //based on the very concise explanation found on SO: https://stackoverflow.com/a/25786793