+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/pki/DistinguishedName.kt
@@ -1,5 +1,3 @@
-@file:OptIn(ExperimentalUnsignedTypes::class)
-
 package at.asitplus.crypto.datatypes.pki
 
 import at.asitplus.crypto.datatypes.asn1.*
@@ -86,16 +84,20 @@ sealed class DistinguishedName : Asn1Encodable<Asn1Set>, Identifiable {
 
         other as DistinguishedName
 
-        return value == other.value
+        if (value != other.value) return false
+        if (oid != other.oid) return false
+
+        return true
     }
 
     override fun hashCode(): Int {
-        return value.hashCode()
+        var result = value.hashCode()
+        result = 31 * result + oid.hashCode()
+        return result
     }
 
     companion object : Asn1Decodable<Asn1Set, DistinguishedName> {
 
-        @OptIn(ExperimentalUnsignedTypes::class)
         @Throws(Asn1Exception::class)
         override fun decodeFromTlv(src: Asn1Set): DistinguishedName = runRethrowing {
             if (src.children.size != 1) throw Asn1StructuralException("Invalid Subject Structure")