+++ datatypes/src/jvmTest/kotlin/PublicKeyTest.kt
@@ -8,6 +8,7 @@ import at.asitplus.crypto.datatypes.io.Base64Strict
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
+import io.kotest.matchers.shouldNotBe
 import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.serialization.encodeToString
@@ -41,9 +42,9 @@ class PublicKeyTest : FreeSpec({
                 println(Json.encodeToString(own))
                 println(own.iosEncoded.encodeToString(Base16()))
                 println(own.encodeToDer().encodeToString(Base16()))
-                println(own.keyId)
+                println(own.didEncoded)
                 own.encodeToDer() shouldBe pubKey.encoded
-                CryptoPublicKey.fromKeyId(own.keyId) shouldBe own
+                CryptoPublicKey.fromDid(own.didEncoded) shouldBe own
                 own.getJcaPublicKey().getOrThrow().encoded shouldBe pubKey.encoded
                 CryptoPublicKey.decodeFromTlv(Asn1Element.parse(own.encodeToDer()) as Asn1Sequence) shouldBe own
             }
@@ -89,7 +90,7 @@ class PublicKeyTest : FreeSpec({
 
                 println(Json.encodeToString(own))
                 println(own.iosEncoded.encodeToString(Base16()))
-                println(own.keyId)
+                println(own.didEncoded)
                 val keyBytes = ((ASN1InputStream(pubKey.encoded).readObject()
                     .toASN1Primitive() as ASN1Sequence).elementAt(1) as DERBitString).bytes
                 own.iosEncoded shouldBe keyBytes //PKCS#1
@@ -105,6 +106,34 @@ class PublicKeyTest : FreeSpec({
 
             pubKey1.hashCode() shouldBe pubKey2.hashCode()
             pubKey1 shouldBe pubKey2
+        }
+    }
+
+    "EC and RSA" - {
+        withData(512, 1024, 2048, 3072, 4096) { rsaBits ->
+            withData(256, 384, 521) { ecBits ->
+                val keyPairEC1 = KeyPairGenerator.getInstance("EC").also { it.initialize(ecBits) }.genKeyPair()
+                val keyPairEC2 = KeyPairGenerator.getInstance("EC").also { it.initialize(ecBits) }.genKeyPair()
+                val keyPairRSA1 = KeyPairGenerator.getInstance("RSA").also { it.initialize(rsaBits) }.genKeyPair()
+                val keyPairRSA2 = KeyPairGenerator.getInstance("RSA").also { it.initialize(rsaBits) }.genKeyPair()
+                val pubKey1 = CryptoPublicKey.decodeFromDer(keyPairEC1.public.encoded)
+                val pubKey2 = CryptoPublicKey.decodeFromDer(keyPairEC2.public.encoded)
+                val pubKey3 = CryptoPublicKey.decodeFromDer(keyPairRSA1.public.encoded)
+                val pubKey4 = CryptoPublicKey.decodeFromDer(keyPairRSA2.public.encoded)
+
+                pubKey1.hashCode() shouldNotBe pubKey2.hashCode()
+                pubKey1.hashCode() shouldNotBe pubKey3.hashCode()
+                pubKey1.hashCode() shouldNotBe pubKey4.hashCode()
+                pubKey3.hashCode() shouldNotBe pubKey4.hashCode()
+                pubKey3.hashCode() shouldNotBe pubKey2.hashCode()
+                pubKey4.hashCode() shouldNotBe pubKey2.hashCode()
+                pubKey1 shouldNotBe pubKey2
+                pubKey1 shouldNotBe pubKey3
+                pubKey1 shouldNotBe pubKey4
+                pubKey3 shouldNotBe pubKey4
+                pubKey3 shouldNotBe pubKey2
+                pubKey4 shouldNotBe pubKey2
+            }
         }
     }
 })